# docs/openapi/CampusHelpATW.openapi.yaml
openapi: 3.0.3
info:
  title: CampusHelpATW API
  version: 0.1.0
  description: |
    CampusHelpATW 微信小程序后端接口（FastAPI 实现基线）。
    实时聊天 WebSocket 端点：
    - `/api/v1/ws/tasks/{task_id}?token={jwt}`
    全局未读通知 WebSocket 端点：
    - `/api/v1/ws/me/notifications?token={jwt}`
    全局通道事件：
    - `unread_snapshot` / `chat_unread`（聊天未读）
    - `task_event`（任务关键节点）
servers:
  - url: http://127.0.0.1:3000
    description: Local
  - url: https://api.example.com
    description: Production (placeholder)
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/reports:
    post:
      tags: [reports]
      summary: 提交举报
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
      responses:
        '200':
          description: 举报结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/reports/mine:
    get:
      tags: [reports]
      summary: 查询我的举报
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 举报列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/admin/reports:
    get:
      tags: [admin]
      summary: 管理端举报列表
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 举报列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
      required: [code, message]
...
    ReportCreateRequest:
      type: object
      required: [target_type, target_id, reason_code]
      properties:
        target_type:
          type: string
        target_id:
          type: integer
        reason_code:
          type: string
        reason_text:
          type: string
...